Moves middleware to /lib directory
I doesn't make sense for the middleware to live inside the /app directory, as this directory is automatically added to the autoload paths of the consuming app. As middleware is only created and added to the stack on initialization (and not re-configured after a hot reload) no changes to the middleware would have any effect on a 'hot' reload anyway.