Adds to_prepare block to Railtie Initializer
The initializer was including the MagicLinks::UrlHelper into ActionController. However, the UrlHelper file is automatically added to the 'hot reloading' list. If the class was updated, the initializer would not be re-run and the included version of UrlHelper would be stale. This was raising the following deprecation warning:

`DEPRECATION WARNING: Initialization autoloaded the constant MagicLinks::UrlHelper.`

I've wrapped the initializer in a `to_prepare` block, so that it get's rerun when the rails application undergoes a 'hot reload'